{"schema":"decodex/commit/1","summary":"Split program intake rendering","authority":"XY-1115"}
